Working Version from Desktop

+++ Spaceport/23-24/Code/Teensy-Based Avionics/src/Radio/APRS/APRSCmdMsg.h
@@ -5,9 +5,9 @@
 
 struct APRSCmdData
 {
-    uint8_t MinutesUntilPowerOn;       // Minutes until the Pi turns on - 0 means turn on now
-    uint8_t MinutesUntilVideoStart;    // Minutes until the Pi starts recording video - 0 means start now
-    uint8_t MinutesUntilDataRecording; // Minutes until the FC starts recording data - 0 means start now
+    int MinutesUntilPowerOn;       // Minutes until the Pi turns on - 0 means turn on now
+    int MinutesUntilVideoStart;    // Minutes until the Pi starts recording video - 0 means start now
+    int MinutesUntilDataRecording; // Minutes until the FC starts recording data - 0 means start now
     bool Launch;                       // Send the launch command to the FC - 0 means don't send
 };
